@@ -14,28 +14,28 @@
 	"devDependencies": {
 		"@feelinglovelynow/svelte-turnstile": "^1.0.0",
 		"@sveltejs/adapter-vercel": "^5.4.1",
-		"@sveltejs/kit": "^2.5.18",
-		"@sveltejs/vite-plugin-svelte": "^3.1.0",
-		"@types/node-fetch": "^2.6.11",
-		"@typescript-eslint/eslint-plugin": "^7.7.0",
-		"@typescript-eslint/parser": "^7.7.0",
-		"eslint": "^9.0.0",
+		"@sveltejs/kit": "^2.7.2",
+		"@sveltejs/vite-plugin-svelte": "4.0.0",
+		"@types/node-fetch": "3.0.2",
+		"@typescript-eslint/eslint-plugin": "8.11.0",
+		"@typescript-eslint/parser": "8.11.0",
+		"eslint": "9.13.0",
 		"eslint-config-prettier": "^9.1.0",
 		"eslint-plugin-svelte3": "^4.0.0",
 		"prettier": "^3.2.5",
 		"prettier-plugin-svelte": "^3.2.3",
 		"sass": "^1.75.0",
-		"svelte": "5.0.0-next.188",
-		"svelte-check": "^3.6.9",
-		"svelte-preprocess": "^5.1.3",
+		"svelte": "5.1.2",
+		"svelte-check": "^4.0.5",
+		"svelte-preprocess": "6.0.3",
 		"tslib": "^2.6.2",
 		"typescript": "^5.4.5",
 		"vite": "^5.2.9"
 	},
 	"type": "module",
 	"dependencies": {
-		"@sveltejs/enhanced-img": "^0.3.0",
-		"@vercel/analytics": "^1.2.2",
+		"@sveltejs/enhanced-img": "^0.3.9",
+		"@vercel/analytics": "1.3.1",
 		"@vercel/speed-insights": "^1.0.10",
 		"node-fetch": "^3.3.2"
 	}
